feat: input label

+++ lib/hexlet_code/tag.rb
@@ -7,7 +7,7 @@ module HexletCode
       br: { default_attrs: {} },
       img: { default_attrs: {} },
       div: { default_attrs: {} },
-      label: { type: "single", default_attrs: {} },
+      label: { type: "double", default_attrs: {} },
       input: { type: "single", default_attrs: { type: "text" } },
       textarea: { type: "double", default_attrs: { cols: 20, rows: 40 } },
       form: { type: "double", default_attrs: { action: "#", method: "post" } }
